pyproject.toml
[project]
name = "heatpump-mcp"
version = "0.1.0"
description = "MCP server for HeatPumpHQ API"
dependencies = [
    "mcp[cli]>=1.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]